# Speed up VarInt encoding, remove emitTilesInOrder option [#98]