--- a/planetiler-core/src/main/java/com/onthegomap/planetiler/util/VarInt.java
+++ b/planetiler-core/src/main/java/com/onthegomap/planetiler/util/VarInt.java
@@ -92,7 +92,7 @@ public class VarInt {
    * Encodes a long integer in a variable-length encoding, 7 bits per byte.
    *
    * @param v            the value to encode
-   * @param outputStream the OutputStream to add the encoded value
+   * @param byteArrayList the OutputStream to add the encoded value
    */
   public static void putVarLong(long v, ByteArrayList byteArrayList) {
     while (true) {