This changes the user and YML-facing configuration by removing the `emitTilesInOrder` option, should we keep that dead option in there for now?

The difference between ordered and unordered MBTiles inserts on a 64 core machine isn't significant: 

ordered:
```
0:53:59 INF - ----------------------------------------
0:53:59 INF -   overall          53m59s cpu:44h37m22s gc:3m39s avg:49.6
0:53:59 INF -   lake_centerlines 3s cpu:10s avg:3.1
0:53:59 INF -     read     2x(29% 0.9s done:2s)
0:53:59 INF -     process 62x(1% 0s wait:2s done:1s)
0:53:59 INF -     write    2x(0% 0s wait:2s done:1s)
0:53:59 INF -   water_polygons   37s cpu:20m36s gc:1s avg:33.6
0:53:59 INF -     read     2x(22% 8s wait:8s done:20s)
0:53:59 INF -     process 62x(49% 18s wait:18s)
0:53:59 INF -     write    2x(63% 23s sys:5s wait:13s)
0:53:59 INF -   natural_earth    20s cpu:34s avg:1.7
0:53:59 INF -     read     2x(34% 7s sys:1s done:13s)
0:53:59 INF -     process 62x(1% 0.1s wait:14s done:6s)
0:53:59 INF -     write    2x(0% 0s wait:15s done:6s)
0:53:59 INF -   osm_pass1        1m54s cpu:1h38m44s gc:7s avg:52
0:53:59 INF -     read     1x(2% 3s sys:1s wait:1m40s done:8s)
0:53:59 INF -     process 63x(77% 1m28s sys:8s block:14s)
0:53:59 INF -   osm_pass2        22m30s cpu:22h2m15s gc:1m24s avg:58.8
0:53:59 INF -     read     1x(0% 4s sys:2s wait:21m44s done:38s)
0:53:59 INF -     process 62x(85% 19m8s sys:2s block:7s wait:23s done:6s)
0:53:59 INF -     write    2x(51% 11m23s sys:2m14s wait:7m30s)
0:53:59 INF -   boundaries       17s cpu:27s avg:1.6
0:53:59 INF -   sort             5m8s cpu:1h15m46s gc:20s avg:14.7
0:53:59 INF -     worker 19x(11% 35s sys:13s wait:2m54s done:13s)
0:53:59 INF -   archive          23m2s cpu:19h18m42s gc:1m46s avg:50.3
0:53:59 INF -     read    2x(20% 4m42s sys:1m32s wait:14m8s done:11s)
0:53:59 INF -     merge   1x(23% 5m21s sys:11s wait:16m11s done:6s)
0:53:59 INF -     encode 62x(74% 17m3s sys:12s wait:4m8s done:6s)
0:53:59 INF -     write   1x(39% 8m55s sys:2m5s wait:12m32s)
0:53:59 INF - ----------------------------------------
```

unordered:
```
0:55:20 INF - ----------------------------------------
0:55:20 INF -   overall          55m20s cpu:44h50m9s gc:3m41s avg:48.6
0:55:20 INF -   lake_centerlines 3s cpu:9s avg:2.9
0:55:20 INF -     read     2x(28% 0.9s done:2s)
0:55:20 INF -     process 62x(1% 0s wait:2s done:1s)
0:55:20 INF -     write    2x(0% 0s wait:2s done:1s)
0:55:20 INF -   water_polygons   38s cpu:20m50s gc:1s avg:32.5
0:55:20 INF -     read     2x(22% 8s wait:9s done:21s)
0:55:20 INF -     process 62x(47% 18s sys:1s wait:19s)
0:55:20 INF -     write    2x(64% 25s sys:5s wait:13s)
0:55:20 INF -   natural_earth    20s cpu:34s avg:1.7
0:55:20 INF -     read     2x(35% 7s sys:1s done:13s)
0:55:20 INF -     process 62x(1% 0.1s wait:14s done:6s)
0:55:20 INF -     write    2x(0% 0s wait:14s done:6s)
0:55:20 INF -   osm_pass1        1m49s cpu:1h34m17s gc:7s avg:51.8
0:55:20 INF -     read     1x(3% 3s sys:2s wait:1m32s done:7s)
0:55:20 INF -     process 63x(77% 1m24s sys:6s block:14s)
0:55:20 INF -   osm_pass2        22m22s cpu:21h54m57s gc:1m23s avg:58.8
0:55:20 INF -     read     1x(0% 3s sys:1s wait:21m41s done:38s)
0:55:20 INF -     process 62x(85% 19m4s sys:2s block:7s wait:29s)
0:55:20 INF -     write    2x(50% 11m14s sys:2m8s wait:7m36s)
0:55:20 INF -   boundaries       17s cpu:25s avg:1.5
0:55:20 INF -   sort             5m11s cpu:1h23m53s gc:19s avg:16.2
0:55:20 INF -     worker 19x(11% 35s sys:13s wait:2m59s done:10s)
0:55:20 INF -   archive          24m34s cpu:19h34m57s gc:1m50s avg:47.8
0:55:20 INF -     read     2x(19% 4m44s sys:1m33s wait:15m42s done:27s)
0:55:20 INF -     merge    1x(22% 5m19s sys:11s wait:17m48s done:18s)
0:55:20 INF -     encoder 62x(70% 17m13s sys:12s wait:5m25s done:8s)
0:55:20 INF -     write    1x(45% 11m6s sys:2m12s wait:12m7s)
0:55:20 INF - ----------------------------------------
```