@msbarry 2023-01-27
Double check for any more references to emit_tiles_in_order, I see one in custtommap readme and planetiler.schema.json.